requirements.txt
pyyaml>=6.0
aiohttp>=3.9.0
aiofiles>=23.0.0
charset-normalizer>=3.3.0
